+++ test/_data/base/foo/file1.js
