+++ test/_data/base/foo/file2.js
