+++ test/_data/base/foo/file3.js
