+++ test/_data/lib/bar/file1.js
