+++ test/_data/lib2/baz/file1.js
