+++ test/_data/lib2/baz/file2.js
