+++ test/bundler/content.js
@@ -2,10 +2,10 @@ var vm = require('vm');
 var assert = require('chai').assert;
 var browserify = require('browserify');
 var convert = require('convert-source-map');
-var bundler = require('../lib/bundler');
+var bundler = require('../../lib/bundler');
 var path = require('path');
 
-var dataDir = path.join(__dirname, 'data');
+var dataDir = path.join(__dirname, '..', '_data');
 var baseDir = path.join(dataDir, 'base');
 var libDir = path.join(dataDir, 'lib');
 
@@ -31,7 +31,7 @@ var executeInVM = function(src) {
   return context;
 };
 
-describe('add content', function() {
+describe('bundler: add content', function() {
 
   it('wrong content', function() {
     var b = browserify();
